backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-jwt-extended==4.7.1
flask-bcrypt==1.0.1
python-dotenv==1.0.1
gunicorn==23.0.0
torch==2.5.1
transformers==4.46.3
Pillow==11.0.0
opencv-python==4.10.0.84
scikit-learn==1.6.1
google-generativeai==0.8.3
playwright==1.49.1
langchain==0.3.17
langchain-google-genai==2.0.9

tavily-python==0.7.21
